add CacheKeyTrait
